+++ src/main/java/group/uchain/project/entity/AllocationForm.java
@@ -3,7 +3,6 @@ import lombok.Data;
 
 import javax.validation.constraints.NotNull;
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Map;
 
 
@@ -22,6 +21,6 @@ public class AllocationForm implements Serializable{
     /**
      * 划分比例以及用户ID
      */
-    Map<Long, BigDecimal> map;
+    Map<Long, Double> map;
 
 }
